+++ src/md2excel/render/CellAppendUtil.java
@@ -1,10 +1,12 @@
-package md2excel;
+package md2excel.render;
 
 import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.CellStyle;
 import org.apache.poi.ss.usermodel.Row;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
+
+import md2excel.excel.MdStyle;
 
 public final class CellAppendUtil {
     private CellAppendUtil() {